refactor: page story 파일 이동
- next build시, .stories.tsx파일까지 포함해서 빌드하는 문제 해결
